Changed some names and moved some packages to make more sense
