+++ alert-common/src/main/java/com/synopsys/integration/alert/common/action/ChannelDistributionTestAction.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.common.descriptor.action;
+package com.synopsys.integration.alert.common.action;
 
 import java.util.Date;
